+++ src/hook/LevelSearchLayer.cpp
@@ -42,20 +42,18 @@ class $modify(RLLevelSearchLayer, LevelSearchLayer) {
                 ->show();
             return;
         }
-        bool disableSearch = Mod::get()->getSettingValue<bool>("disableRlSearchLayer"); // please work no. 2
-                CCLayer* layer = nullptr; 
-        
-                if (disableSearch) {
-                    layer = RLMenuLayer::create();
-                } else {
-                    layer = RLSearchLayer::create();
-                }
-        
-                if (layer) {
-                    auto scene = CCScene::create();
-                    scene->addChild(layer);
-                    auto transitionFade = CCTransitionFade::create(0.5f, scene);
-                    CCDirector::sharedDirector()->pushScene(transitionFade);
-                }
-            }
-        };
+        bool disableSearch = Mod::get()->getSettingValue<bool>("disableRLSearchLayer");  // please work no. 2
+        CCLayer* layer = nullptr;
+        if (disableSearch) {
+            layer = RLMenuLayer::create();
+        } else {
+            layer = RLSearchLayer::create();
+        }
+        if (layer) {
+            auto scene = CCScene::create();
+            scene->addChild(layer);
+            auto transitionFade = CCTransitionFade::create(0.5f, scene);
+            CCDirector::sharedDirector()->pushScene(transitionFade);
+        }
+    }
+};