+++ src/layer/RLSpireLayer.cpp
@@ -108,6 +108,7 @@ void RLSpireLayer::onSpireClick(CCObject* sender) {
                 break;
             case 3:
                 response = "I heard <cp>The Oracle</c> is pretty <cl>knowledgeable</c>. Maybe you should ask it about this <cf>Spire</c>.";
+                m_indexDia++;
                 break;
             default:
                 m_indexDia = 0;