+++ changelog.md
@@ -1,6 +1,12 @@
 # v1.0.14
 
-- Changed the submission queue button sprite
+- Changed the submission queue button sprite.
+- Changed the colors of the moderators/boosters roles and credits.
+- Add the missing Leaderboard Admin Badge on the comment cell.
+- Added Discord Username in the claim supporter badge popup.
+- Moderators Notes shows in multiple lines instead of single line.
+- Fixed issue where the Rubies value goes to the negative.
+- Moderator's Username are now clickable in the Moderators Notes Popup.
 
 # v1.0.13